+++ src/main/java/sonia/scm/jenkins/JenkinsRepositoryHookHandler.java
@@ -30,27 +30,22 @@ import com.cloudogu.scm.el.env.ImmutableEncodedRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import sonia.scm.net.ahc.AdvancedHttpClient;
-import sonia.scm.net.ahc.AdvancedHttpRequest;
 import sonia.scm.net.ahc.AdvancedHttpRequestWithBody;
-import sonia.scm.net.ahc.AdvancedHttpResponse;
-import sonia.scm.net.ahc.BaseHttpRequest;
 import sonia.scm.repository.Changeset;
 import sonia.scm.repository.RepositoryHookEvent;
 import sonia.scm.util.HttpUtil;
-import sonia.scm.util.IOUtil;
 import sonia.scm.util.Util;
 
 import javax.inject.Provider;
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.MalformedURLException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+
+import static sonia.scm.jenkins.HeaderAppenders.appendAuthenticationHeader;
+import static sonia.scm.jenkins.HeaderAppenders.appendCsrfCrumbHeader;
+import static sonia.scm.jenkins.Urls.escape;
 
 /**
  * @author Sebastian Sdorra
@@ -107,7 +102,7 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler {
     String url = createBaseUrl(configuration);
 
     //J-
-    // url encode proejct name, see http://goo.gl/v8Rond
+    // url encode project name, see http://goo.gl/v8Rond
     return escape(
       url.concat("job/")
         .concat(configuration.getProject())
@@ -117,44 +112,12 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler {
   }
 
   /**
-   * Escapes a complete url. The method is visible to the package for testing.
-   *
-   * @param urlString url
-   * @return escaped url
-   */
-  String escape(String urlString) {
-    String escapedUrl = urlString;
-
-    try {
-      URL url = new URL(urlString);
-      //J-
-      URI uri = new URI(
-        url.getProtocol(),
-        url.getUserInfo(),
-        url.getHost(),
-        url.getPort(),
-        url.getPath(),
-        url.getQuery(),
-        url.getRef()
-      );
-      //J+
-
-      escapedUrl = uri.toString();
-    } catch (URISyntaxException | MalformedURLException ex) {
-      logger.warn("could not escaped url", ex);
-    }
-
-    return escapedUrl;
-  }
-
-  /**
    * Handles the repository event. Checks the configuration for the repository
    * and calls the jenkins server.
    *
    * @param configuration jenkins configuration
    */
   private void handleRepositoryEvent(JenkinsConfiguration configuration, RepositoryHookEvent event) {
-
     String url = createUrl(configuration);
 
     if (logger.isInfoEnabled()) {
@@ -165,75 +128,6 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler {
       sendRequest(configuration, event, url);
     } catch (IOException ex) {
       logger.error("could not send request to jenkins", ex);
-    }
-  }
-
-  private String createCrumbUrl(JenkinsConfiguration configuration) {
-    return escape(createBaseUrl(configuration).concat("crumbIssuer/api/xml"));
-  }
-
-  private CsrfCrumb parseCsrfCrumbResponse(AdvancedHttpResponse response) throws IOException {
-    CsrfCrumb csrfCrumb;
-    InputStream content = null;
-    try {
-      content = response.contentAsStream();
-      csrfCrumb = CsrfCrumbParser.parse(content);
-    } finally {
-      IOUtil.close(content);
-    }
-    return csrfCrumb;
-  }
-
-  private CsrfCrumb getJenkinsCsrfCrumb(JenkinsConfiguration configuration, AdvancedHttpClient client) {
-    String url = createCrumbUrl(configuration);
-    logger.debug("fetch csrf crumb from {}", url);
-
-    AdvancedHttpRequest request = client.get(url).spanKind("Jenkins");
-
-    appendAuthenticationHeader(configuration, request);
-
-    CsrfCrumb crumb = null;
-    try {
-      AdvancedHttpResponse response = request.request();
-      int sc = response.getStatus();
-      if (sc != 200) {
-        logger.warn("jenkins crumb endpoint returned status code {}", sc);
-      } else {
-        crumb = parseCsrfCrumbResponse(response);
-      }
-    } catch (IOException ex) {
-      logger.warn("failed to fetch csrf crumb", ex);
-    }
-
-    return crumb;
-  }
-
-  private void appendCsrfCrumb(JenkinsConfiguration configuration, AdvancedHttpClient client, BaseHttpRequest request) {
-    if (configuration.isCsrf()) {
-      CsrfCrumb crumb = getJenkinsCsrfCrumb(configuration, client);
-      if (crumb != null) {
-        logger.debug("add csrf crumb to api request");
-        request.header(crumb.getCrumbRequestField(), crumb.getCrumb());
-      }
-    } else {
-      logger.debug("csrf protection is disabled, skipping crumb");
-    }
-  }
-
-  private void appendAuthenticationHeader(JenkinsConfiguration configuration, BaseHttpRequest request) {
-    // check for authentication parameters
-    String username = configuration.getUsername();
-    String apiToken = configuration.getApiToken();
-
-    if (Util.isNotEmpty(username) && Util.isNotEmpty(apiToken)) {
-      if (logger.isDebugEnabled()) {
-        logger.debug("added authentication for user {}", username);
-      }
-
-      // add basic authentication header
-      request.basicAuth(username, apiToken);
-    } else if (logger.isDebugEnabled()) {
-      logger.debug("skip authentication. username or api token is empty");
     }
   }
 
@@ -265,8 +159,10 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler {
       logger.debug("no project token is available");
     }
 
-    appendAuthenticationHeader(configuration, request);
-    appendCsrfCrumb(configuration, httpClient, request);
+    appendAuthenticationHeader(request, configuration.getUsername(), configuration.getApiToken());
+    if (configuration.isCsrf()) {
+      appendCsrfCrumbHeader(httpClient, request, configuration.getUrl(), configuration.getUsername(), configuration.getApiToken());
+    }
 
     /**
      * execute the http post request with the http client and
@@ -313,7 +209,7 @@ public class JenkinsRepositoryHookHandler implements JenkinsHookHandler {
           .append(HttpUtil.encode(HttpUtil.decode(elParser.parse(parameter.getValue()).evaluate(env))))
           .append("&")
       );
-      url = builder.toString().substring(0, builder.toString().length() - 1);
+      url = builder.substring(0, builder.length() - 1);
     }
 
     return url;