+++ src/main/js/GlobalJenkinsConfigurationForm.tsx
@@ -31,6 +31,9 @@ type GlobalConfiguration = {
   disableMercurialTrigger: boolean;
   disableGitTrigger: boolean;
   disableEventTrigger: boolean;
+  disableSubversionTrigger: boolean;
+  username?: string;
+  apiToken?: string;
 };
 
 type Props = WithTranslation & {
@@ -105,11 +108,36 @@ class GlobalJenkinsConfigurationForm extends React.Component<Props, State> {
           onChange={this.valueChangeHandler}
         />
         <Checkbox
+          name={"disableSubversionTrigger"}
+          label={t("scm-jenkins-plugin.global.form.disableSubversionTrigger")}
+          helpText={t("scm-jenkins-plugin.global.form.disableSubversionTriggerHelp")}
+          checked={this.state.disableSubversionTrigger}
+          disabled={readOnly}
+          onChange={this.valueChangeHandler}
+        />
+        <Checkbox
           name={"disableEventTrigger"}
           label={t("scm-jenkins-plugin.global.form.disableEventTrigger")}
           helpText={t("scm-jenkins-plugin.global.form.disableEventTriggerHelp")}
           checked={this.state.disableEventTrigger}
           disabled={readOnly}
+          onChange={this.valueChangeHandler}
+        />
+        <InputField
+          name={"username"}
+          label={t("scm-jenkins-plugin.global.form.username")}
+          helpText={t("scm-jenkins-plugin.global.form.usernameHelp")}
+          disabled={readOnly || this.state.disableSubversionTrigger}
+          value={this.state.username}
+          onChange={this.valueChangeHandler}
+        />
+        <InputField
+          name={"apiToken"}
+          type="password"
+          label={t("scm-jenkins-plugin.global.form.apiToken")}
+          helpText={t("scm-jenkins-plugin.global.form.apiTokenHelp")}
+          disabled={readOnly || this.state.disableSubversionTrigger}
+          value={this.state.apiToken}
           onChange={this.valueChangeHandler}
         />
       </>