+++ src/test/java/sonia/scm/jenkins/JenkinsGlobalHookHandlerTest.java
@@ -21,62 +21,133 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package sonia.scm.jenkins;
 
-//~--- non-JDK imports --------------------------------------------------------
+import com.google.inject.Provider;
+import com.google.inject.util.Providers;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Answers;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import sonia.scm.net.ahc.AdvancedHttpClient;
+import sonia.scm.net.ahc.AdvancedHttpRequest;
+import sonia.scm.net.ahc.AdvancedHttpResponse;
+import sonia.scm.repository.Repository;
+import sonia.scm.repository.RepositoryHookEvent;
+import sonia.scm.repository.RepositoryHookType;
+import sonia.scm.repository.RepositoryTestData;
+import sonia.scm.repository.api.HookContext;
+import sonia.scm.repository.api.RepositoryService;
+import sonia.scm.repository.api.RepositoryServiceFactory;
 
-import org.junit.Test;
+import java.io.IOException;
 
-import static org.junit.Assert.*;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.lenient;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
-/**
- *
- * @author Sebastian Sdorra
- */
-public class JenkinsGlobalHookHandlerTest
-{
+@ExtendWith(MockitoExtension.class)
+class JenkinsGlobalHookHandlerTest {
 
-  /**
-   * Method description
-   *
-   */
+  @Mock
+  private RepositoryServiceFactory serviceFactory;
+  @Mock
+  private RepositoryService repositoryService;
+  @Mock
+  private GlobalJenkinsConfiguration config;
+  @Mock
+  private AdvancedHttpClient advancedHttpClient;
+  @Mock(answer = Answers.RETURNS_SELF)
+  private AdvancedHttpRequest request;
+  @Mock
+  private AdvancedHttpResponse response;
+  @Mock(answer = Answers.RETURNS_DEEP_STUBS)
+  private HookContext hookContext;
+
+  private JenkinsGlobalHookHandler handler;
+
+  @BeforeEach
+  void initClient() {
+    Provider<AdvancedHttpClient> httpClientProvider = Providers.of(advancedHttpClient);
+    handler = new JenkinsGlobalHookHandler(httpClientProvider, config, serviceFactory);
+  }
+
   @Test
-  public void testFixUrl()
-  {
-    //J-
-    assertEquals(
-      "http://www.scm-manager.org/scm/hg", 
-      JenkinsGlobalHookHandler.fixUrl("http://www.scm-manager.org/scm/hg")
-    );
-    assertEquals(
-      "http://www.scm-manager.org/scm/hg", 
-      JenkinsGlobalHookHandler.fixUrl("http://www.scm-manager.org/scm//hg")
-    );
-    assertEquals(
-      "http://www.scm-manager.org/scm/hg", 
-      JenkinsGlobalHookHandler.fixUrl("http://www.scm-manager.org//scm//hg")
-    );
-    assertEquals(
-      "https://www.scm-manager.org/scm/hg", 
-      JenkinsGlobalHookHandler.fixUrl("https://www.scm-manager.org//scm//hg")
-    );
-    assertEquals(
-      "https://www.scm-manager.org/scm/hg", 
-      JenkinsGlobalHookHandler.fixUrl("https://www.scm-manager.org/scm/hg")
-    );
-    assertEquals(
-      "/scm/hg", 
-      JenkinsGlobalHookHandler.fixUrl("/scm/hg")
-    );
-    assertEquals(
-      "/scm/hg", 
-      JenkinsGlobalHookHandler.fixUrl("//scm/hg")
-    );
-    assertEquals(
-      "/scm/hg", 
-      JenkinsGlobalHookHandler.fixUrl("//scm//hg")
-    );
-    //J+
+  void shouldNotSendRequestIfConfigInvalid() {
+    Repository repository = RepositoryTestData.createHeartOfGold();
+    when(config.isValid()).thenReturn(false);
+
+    handler.sendRequest(new RepositoryHookEvent(hookContext, repository, RepositoryHookType.POST_RECEIVE));
+
+    verify(advancedHttpClient, never()).get(anyString());
+  }
+
+  @Nested
+  class WithValidConfig {
+    @BeforeEach
+    void initValidConfig() {
+      when(config.isValid()).thenReturn(true);
+      lenient().when(config.getUrl()).thenReturn("jenkins.io/scm/");
+    }
+
+    @Test
+    void shouldNotSendRequestForMercurialRepoIfTriggerIsDisabled() {
+      when(config.isDisableMercurialTrigger()).thenReturn(true);
+      Repository repository = RepositoryTestData.createHeartOfGold();
+      repository.setType("hg");
+
+      handler.sendRequest(new RepositoryHookEvent(hookContext, repository, RepositoryHookType.POST_RECEIVE));
+
+      verify(advancedHttpClient, never()).get(anyString());
+    }
+
+    @Test
+    void shouldSendRequestForMercurialRepoIfTriggerEnabled() throws IOException {
+      Repository repository = RepositoryTestData.createHeartOfGold();
+      repository.setType("hg");
+
+      when(config.isDisableMercurialTrigger()).thenReturn(false);
+      when(serviceFactory.create(repository)).thenReturn(repositoryService);
+      mockHttpClient(repository);
+
+      handler.sendRequest(new RepositoryHookEvent(hookContext, repository, RepositoryHookType.POST_RECEIVE));
+
+      verify(advancedHttpClient).get("jenkins.io/scm/mercurial/notifyCommit");
+    }
+
+    @Test
+    void shouldNotSendRequestForGitRepoIfTriggerIsDisabled() {
+      when(config.isDisableGitTrigger()).thenReturn(true);
+      Repository repository = RepositoryTestData.createHeartOfGold();
+      repository.setType("git");
+
+      handler.sendRequest(new RepositoryHookEvent(hookContext, repository, RepositoryHookType.POST_RECEIVE));
+
+      verify(advancedHttpClient, never()).get(anyString());
+    }
+
+    @Test
+    void shouldSendRequestForGitRepoIfTriggerIsEnabled() throws IOException {
+      Repository repository = RepositoryTestData.createHeartOfGold();
+      repository.setType("git");
+      when(config.isDisableGitTrigger()).thenReturn(false);
+      mockHttpClient(repository);
+
+      handler.sendRequest(new RepositoryHookEvent(hookContext, repository, RepositoryHookType.POST_RECEIVE));
+
+      verify(advancedHttpClient).get("jenkins.io/scm/git/notifyCommit");
+    }
+
+    private void mockHttpClient(Repository repository) throws IOException {
+      when(serviceFactory.create(repository)).thenReturn(repositoryService);
+      when(advancedHttpClient.get(anyString())).thenReturn(request);
+      when(request.request()).thenReturn(response);
+      when(response.getStatus()).thenReturn(200);
+    }
   }
 }