+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Send trigger events to the global jenkins instance to trigger updates for namespace navigator items ([#13](https://github.com/scm-manager/scm-jenkins-plugin/pull/13))
+- Add SVN global hook handler ([#14](https://github.com/scm-manager/scm-jenkins-plugin/pull/14))
 
 ### Changed
 - Set span kind for http requests (for Trace Monitor)